requirements.txt
flask
nltk
gunicorn
google-generativeai
python-dotenv
requests